naprawione nadirPoint i chyba ED

+++ src/ParetoFrontsGenerator.java
@@ -135,9 +135,9 @@ class ParetoFrontsGenerator {
         ArrayList<Individual> paretoFront = paretoFronts.get(0);
         paretoFront.sort(new ObjectiveFrontComparator());
         for (Individual ind : paretoFront) {
-            sumED += Math.sqrt((long) (int) ((ind.getFitnessTime() - ideal.x)
-                    * (ind.getFitnessTime() - ideal.x) + (ind.getFitnessWage() - ideal.y)
-                    * (ind.getFitnessWage() - ideal.y)));
+            sumED += Math.sqrt((long) (((ind.getFitnessTime() - ideal.x)
+                    * (ind.getFitnessTime() - ideal.x)) + ((ind.getFitnessWage() - ideal.y)
+                    * (ind.getFitnessWage() - ideal.y))));
         }
 
         return (sumED / paretoFront.size()) + "";
